feat: 🎸 [ir_visualizer]Render graphs in one file
Previously, ir_visualizer always generates seperate files for multiple
graphs(GNetworks), considering muticore case of NPU. However, for our
models, the body of if or loop are also stored in GNetworks. So it's
not a good idea to generate a file for each GNetwork. Otherwise there
will be lots of output files if there are many if or loop Op, and it's
very inconvenient to recognize the relation of them.
So we change to place all GNetworks in one file. And then, we can try
to embed the body of if or loop into the original graph, to make the
visualization result more readable.

Signed-off-by: jiwei-yu <jiwei.yu@samsung.com>