feat: 🎸 Add json submodule & parse util functions
1. Add json submodule. NPUCompiler uses amalgamated json files,
which are too large to import directly (code-format requirement).
And the amalgamated cpp file is generated by python script, which
means it cannot be imported as an ExternalProject in cmake.
Whatsmore, these files are generated by json in a history version
which we are not aware of. So we add json as a submodule and read
in from PIM library.
2. Pass service is supported for json file parsing: pass configs
will be stored in a json file. Add util functions at first.

Signed-off-by: jiwei-yu <jiwei.yu@samsung.com>